Add files via upload
Sentinel Tracker is a human-following robot designed to autonomously track and follow a specific person using sensor-based navigation. The robot leverages ultrasonic sensors, IR sensors, and a microcontroller to detect and follow a target, making it ideal for applications such as assistance for disabled individuals, automated load carrying, and surveillance.

Components Used
1) Arduino Uno 
2) Motor Driver Shield 
3) Wheels (4x)
4) TT Gear Motor (4x) 
5) Servo Motor 
6) Ultrasonic Sensor 
7) Infrared Sensor (2x) 
) 18650 Li-on Battery (2x) 
7) 18650 Battery Holder 
8) Male and Female Jumper wire 
9) Acrylic Sheet
10) DC Power Switch 

Working Principle
Sensing the Target: The ultrasonic and IR sensors detect the presence and movement of a person.
Data Processing: The microcontroller processes sensor data to determine movement.
Motor Control: Based on the processed data, the L298N driver controls the DC motors to move the robot in the required direction.
Obstacle Avoidance: If an obstacle is detected, the robot adjusts its path to avoid collisions.